backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
databricks-sql-connector==3.5.0
pydantic==2.11.7
python-dotenv==1.0.1
